moved stuff around for more coherence
