+++ models/games/card_games/variances/president/__init__.py
@@ -5,8 +5,8 @@ Project: President-Game
 IDE: PyCharm
 Creation-date: 11/25/22
 """
-from .president import PresidentGame
 
+from .president_game import PresidentGame
 
 if __name__ == '__main__':
     pass
