+++ models/games/card_games/variances/president/president_rankings.py
